requirements.txt
langchain
wikipedia
langchain-ollama
langchain-community
langchain-anthropic
python-dotenv
pydantic
pypdf
PyPDF2
python-docx
sentence-transformers
numpy
reportlab